[TASK] Initial commit
This is a proof of concept for a new TypoScript Content Object IMAGE_GALLERY, which is similar to the Content Object FILES, but with more options for rendering an image gallery.

It is developed as a standalone extension.

This patch adds the same code as FILES, but in a more structured way.