[FEATURE] Rendering HTML5
This patch offers a lot of new features, and renders like the old pi1 class of css_styled_content. It's only HTML5 and is lacking the following settings:

- colSpace
- border
- borderThick
- borderSpace
- noRows
- textMargin
- image_compression
- image_effects
- imagecaption_position

The calculation of the images done without taking these settings into account. Some of these settings IMHO are not for editors, but for styling.